refactor: controller, model 추가
